requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
tensorboard>=2.10.0
diffusers>=0.20.0
transformers>=4.20.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0
wandb>=0.15.0
python-dotenv>=0.19.0
datasets>=2.0.0 